Update: translator のリファクタリング (#41)
