+++ src/llm_translator.py
@@ -1,16 +1,8 @@
-import os
 import json
 
+from llm import LLM
 from bunkai import Bunkai
-from langchain.chat_models import ChatOpenAI
-from langchain import LLMChain
-from langchain.prompts.chat import (
-    ChatPromptTemplate,
-    SystemMessagePromptTemplate,
-    HumanMessagePromptTemplate,
-)
-from langchain.callbacks import get_openai_callback
-from prompt import SYSTEM_TEMPLATE, TEXT_OUTPUT_FORMAT, TABLE_OUTPUT_FOTMAT, HUMAN_TEMPLATE, EXAMPLE_EN_TO_JA, EXAMPLE_JA_TO_EN, TEXT_EXAMPLE
+from translation_prompt import SYSTEM_TEMPLATE, TEXT_OUTPUT_FORMAT, TABLE_OUTPUT_FOTMAT, HUMAN_TEMPLATE, EXAMPLE_EN_TO_JA, EXAMPLE_JA_TO_EN, TEXT_EXAMPLE
 from const import EN, JA
 
 
@@ -67,44 +59,10 @@ class SplitedSentence:
         return str(self.texts)
 
 
-class Translator:
+class Translator(LLM):
     def __init__(self, debug: bool=False):
-        if not os.environ.get("OPENAI_API_KEY"):
-            raise ValueError("OPENAI_API_KEY is not set.")
+        super().__init__(debug=debug)
         self.japaneses_splitter = Bunkai()
-        self.debug = debug
-        self.BEGIN_JSON_FORMAT = "```json\n"
-        self.END_JSON_FORMAT = "\n```"
-
-    def _get_llm(self, model: str="gpt-3.5-turbo", temperature: float=0.0, max_tokens: int=2000) -> ChatOpenAI:
-        return ChatOpenAI(
-            model=model,
-            temperature=temperature,
-            max_tokens=max_tokens,
-        )
-
-    def _make_llm_chain(self, llm, system_template, human_template) -> LLMChain:
-        system_message_prompt = SystemMessagePromptTemplate.from_template(system_template)
-        human_message_prompt = HumanMessagePromptTemplate.from_template(human_template)
-        chat_prompt = ChatPromptTemplate.from_messages([system_message_prompt, human_message_prompt])
-        chain = LLMChain(llm=llm, prompt=chat_prompt)
-        return chain
-
-    def _run_llm_chain(self, chain: LLMChain, **kwargs) -> tuple:
-        with get_openai_callback() as _callback:
-            response = chain.run(kwargs)
-            cost = _callback.total_cost
-            tokens = _callback.total_tokens
-        try:
-            if self.debug:
-                print(response)
-            if response.startswith(self.BEGIN_JSON_FORMAT) and response.endswith(self.END_JSON_FORMAT):
-                response = response[len(self.BEGIN_JSON_FORMAT): -len(self.END_JSON_FORMAT)]
-            result = json.loads(response, strict=False)
-        except json.decoder.JSONDecodeError:
-            return {}, cost, tokens
-
-        return result, cost, tokens
 
     def translate_by_sentence(self, source_language, target_language, text, model, temperature):
         """
