+++ src/translation_prompt.py
@@ -16,27 +16,27 @@ TABLE_OUTPUT_FOTMAT = ("- The outcome should be in JSON format.\n"
                        "- The translated results are stored in an array for each sentence, where the string \"translated_texts\" is used as the key, and the translated text corresponding to the key is stored for each sentence.\n")
 
 EXAMPLE_EN_TO_JA = ("```input (array)\n"
-                    "[\"Hello, how are you?\", \"I'm fine, thank you.\", \"Hello, how are you?\"]"
+                    "[\"Hello, how are you?\", \"I'm fine, thank you.\", \"Hello, world!\"]"
                     "```\n"
                     "``` output (json)\n"
                     "{\n"
                     "  \"translated_texts\": [\n"
                     "    \"こんにちは、お元気ですか？\",\n"
                     "    \"ありがとう。\",\n"
-                    "    \"こんにちは、お元気ですか？\"\n"
+                    "    \"こんにちは、世界！\"\n"
                     "  ]\n"
                     "}\n"
                     "```\n")
 
 EXAMPLE_JA_TO_EN = ("```input (array)\n"
-                    "[\"こんにちは、お元気ですか？\", \"ありがとう。\", \"こんにちは、お元気ですか？\"]"
+                    "[\"こんにちは、お元気ですか？\", \"ありがとう。\", \"こんにちは、世界！\"]"
                     "```\n"
                     "``` output (json)\n"
                     "{\n"
                     "  \"translated_texts\": [\n"
                     "    \"Hello, how are you?\",\n"
                     "    \"I'm fine, thank you.\",\n"
-                    "    \"Hello, how are you?\"\n"
+                    "    \"Hello, world!\"\n"
                     "  ]\n"
                     "}\n"
                     "```\n")
